squash some compiler warnings on debug printfs by casting to int

+++ sys/ufs/lfs/lfs_alloc.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: lfs_alloc.c,v 1.22 1999/06/15 22:25:41 perseant Exp $	*/
+/*	$NetBSD: lfs_alloc.c,v 1.23 1999/06/17 22:22:41 tls Exp $	*/
 
 /*-
  * Copyright (c) 1999 The NetBSD Foundation, Inc.
@@ -352,7 +352,7 @@ lfs_vfree(v)
 		LFS_SEGENTRY(sup, fs, datosn(fs, old_iaddr), bp);
 #ifdef DIAGNOSTIC
 		if (sup->su_nbytes < DINODE_SIZE) {
-			printf("lfs_vfree: negative byte count (segment %d short by %d)\n", datosn(fs, old_iaddr), DINODE_SIZE - sup->su_nbytes);
+			printf("lfs_vfree: negative byte count (segment %d short by %d)\n", datosn(fs, old_iaddr), (int)DINODE_SIZE - sup->su_nbytes);
 			panic("lfs_vfree: negative byte count");
 			sup->su_nbytes = DINODE_SIZE;
 		}
